openstack-ardana: remove SLES SDK usage
The CI shouldn't rely on `gcc` and `python-devel` to be installed on
the deployer node to build packages at runtime, it should always use
packages coming from OBS or IBS instead.

This commit removes SLES SDK repositories to enforce this policy and
uses the `virtualenv_site_packages` virtualenv option to allow pip
to reuse system packages.